Consolidated shared code
