+++ BthPS3.sln
@@ -40,6 +40,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BthPS3Installer", "BthPS3Installer\BthPS3Installer.csproj", "{85268F54-2547-41E1-932A-5D1341D7A11E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nefarius.BthPS3.Shared", "shared\Nefarius.BthPS3.Shared.csproj", "{105DE3D0-E7CA-4600-9600-745468742AA1}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -140,6 +142,18 @@ Global
 		{85268F54-2547-41E1-932A-5D1341D7A11E}.Release|ARM64.Build.0 = Release|Any CPU
 		{85268F54-2547-41E1-932A-5D1341D7A11E}.Release|x64.ActiveCfg = Release|Any CPU
 		{85268F54-2547-41E1-932A-5D1341D7A11E}.Release|x64.Build.0 = Release|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Debug|x64.Build.0 = Debug|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Release|ARM64.Build.0 = Release|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Release|x64.ActiveCfg = Release|Any CPU
+		{105DE3D0-E7CA-4600-9600-745468742AA1}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -155,6 +169,7 @@ Global
 		{81C9766C-7FD8-473D-882D-85FD89EE3C4C} = {2BD82ADC-39FF-4E32-BEBB-B0DD956FBDE4}
 		{93A044C1-9491-4A17-B410-D144DBBD8F44} = {2BD82ADC-39FF-4E32-BEBB-B0DD956FBDE4}
 		{85268F54-2547-41E1-932A-5D1341D7A11E} = {2BD82ADC-39FF-4E32-BEBB-B0DD956FBDE4}
+		{105DE3D0-E7CA-4600-9600-745468742AA1} = {51B50044-0861-4230-8832-80E1F44D472E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9602E54E-A4C1-432E-9E5F-E4C5FF32985C}
