+++ BthPS3Installer/BthPS3Installer.csproj
@@ -30,9 +30,13 @@
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="6.0.0" />
     <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.4" />
-    <PackageReference Include="Nefarius.Utilities.Bluetooth" Version="1.5.0" />
+    <PackageReference Include="Nefarius.Utilities.Bluetooth" Version="1.6.0" />
     <PackageReference Include="Nefarius.Utilities.DeviceManagement" Version="3.23.0" />
     <PackageReference Include="WixSharp_wix4" Version="2.3.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\shared\Nefarius.BthPS3.Shared.csproj" />
   </ItemGroup>
 
   <ItemGroup>
